backend/requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
joblib==1.3.2
Markdown==3.5.1
mysqlclient==2.2.0
nltk==3.8.1
python-decouple==3.8
pytz==2023.3.post1
regex==2023.10.3
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
